Adjusted text for "100% User Owned" Section
I added the following text "Essentially, DropTown enables you to buy and sell e-businesses through the power of NFTs. That means if you’ve built your site into a profitable business it could constitute a higher asking price!" to the 2nd paragraph as we felt it was more clear to the readers.